+++ src/main/java/duke/DukeException.java
@@ -1,19 +1,18 @@
-package Duke;
+package duke;
 
 public class DukeException extends Exception {
     private String msg;
 
-//    public DukeException(String msg, Throwable cause) {
-//        super(msg, cause);
-//    }
-//
+    /**
+     * Constructor for a duke exception
+     * @param msg a text to be printed out
+     */
     public DukeException(String msg) {
         this.msg = msg;
     }
 
     @Override
     public String toString() {
-        //super(cause);
         if (msg.equals("file not found")) {
             return "no databse found! pls try again ^__^";
         } else if (msg.equals("command not found")) {
