+++ src/main/java/duke/ReadFile.java
@@ -1,4 +1,4 @@
-package Duke;
+package duke;
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.time.LocalDate;
