+++ src/main/java/duke/Todo.java
@@ -1,4 +1,4 @@
-package Duke;
+package duke;
 
 public class Todo extends Task {
