+++ src/test/java/DukeTest.java
@@ -1,6 +1,6 @@
-import Duke.Deadline;
-import Duke.Event;
-import Duke.Task;
+import duke.Deadline;
+import duke.Event;
+import duke.Task;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
